server/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
mysql-connector-python==8.2.0 